Rename API folders to match their exported values
