+++ packages/admin-ui-extensions-react/src/extension-api/ContainerApi/examples/container.example.tsx
