+++ packages/admin-ui-extensions-react/src/extension-api/LocaleApi/examples/locale.example.tsx
