+++ packages/admin-ui-extensions-react/src/extension-api/SessionTokenApi/examples/sessionToken.example.tsx
