+++ packages/admin-ui-extensions-react/src/extension-api/SessionTokenApi/index.ts
