+++ packages/admin-ui-extensions-react/src/extension-api/index.ts
@@ -1,7 +1,7 @@
 export {useExtensionApi} from './utils';
-export * from './layout';
-export * from './locale';
-export * from './container';
-export * from './sessionToken';
-export * from './toast';
-export * from './data';
+export * from './LayoutApi';
+export * from './LocaleApi';
+export * from './ContainerApi';
+export * from './SessionTokenApi';
+export * from './ToastApi';
+export * from './DataApi';
