+++ packages/admin-ui-extensions/src/extension-api/LocaleApi/examples/locale.example.ts
