+++ packages/admin-ui-extensions/src/extension-api/SessionTokenApi/examples/sessionToken.example.ts
