+++ packages/admin-ui-extensions/src/extension-points/types.ts
@@ -1,9 +1,9 @@
 import {RemoteRoot, RemoteComponentType} from '@remote-ui/core';
 
-import type {LayoutApi} from '../extension-api/layout';
-import type {LocaleApi} from '../extension-api/locale';
-import type {ToastApi} from '../extension-api/toast';
-import type {SessionTokenApi} from '../extension-api/sessionToken';
+import type {LayoutApi} from '../extension-api/LayoutApi';
+import type {LocaleApi} from '../extension-api/LocaleApi';
+import type {ToastApi} from '../extension-api/ToastApi';
+import type {SessionTokenApi} from '../extension-api/SessionTokenApi';
 
 export type ExtensionResult = Record<string, never> | void;
